Refactoring.

+++ LiveRiotSocial/FHSTwitterEngine.m
@@ -1244,16 +1244,16 @@ id removeNull(id rootObject) {
         return [NSArray arrayWithObjects:initialString, nil];
     }
     
-    int offset = 0;
+    NSInteger offset = 0;
     int remainder = fmod(array.count, 100);
-    int numberOfStrings = (array.count-remainder)/100;
+    int numberOfStrings = (int)(array.count-remainder)/100;
     
     NSMutableArray *reqStrs = [NSMutableArray array];
     
     for (int i = 1; i <= numberOfStrings; ++i) {
         NSString *ninetyNinththItem = (NSString *)[array objectAtIndex:(i*100)-1];
         NSRange range = [initialString rangeOfString:ninetyNinththItem];
-        int endOffset = range.location+range.length;
+        NSInteger endOffset = range.location + range.length;
         NSRange rangeOfAString = NSMakeRange(offset, endOffset-offset);
         offset = endOffset;
         NSString *endResult = [initialString fhs_stringWithRange:rangeOfAString];
@@ -1982,7 +1982,7 @@ static char const Encode[] = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwx
 @implementation NSData (Base64)
 
 - (NSString *)base64Encode {
-    int outLength = ((((self.length*4)/3)/4)*4)+(((self.length*4)/3)%4?4:0);
+    NSInteger outLength = ((((self.length*4)/3)/4)*4)+(((self.length*4)/3)%4?4:0);
     const char *inputBuffer = self.bytes;
     char *outputBuffer = malloc(outLength+1);
     outputBuffer[outLength] = 0;
